Updated login, reg, stat page

+++ client/src/pages/login/login.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import Axios from 'axios'
 import { useState } from 'react';
+import firebase from '../../firebase';
 
 export default function Login() {
 
@@ -11,17 +12,28 @@ export default function Login() {
     const login = (e) => {
         e.preventDefault();
 
-        Axios.post("http://localhost:3001/login", {
-            NID, phone
-        }).then((response) => {
-            if (response.data.message) {
-                setLoginErr(response.data.message);
-            }else {
-                localStorage.setItem("loginStat", "yes");
-                localStorage.setItem("loginCred", JSON.stringify(response.data[0]));
-                console.log(response.data[0]);
-            }
-        });
+        let recaptcha = new firebase.auth.RecaptchaVerifier('recaptcha');
+        let number = phone;
+        firebase.auth().signInWithPhoneNumber(number, recaptcha).then(function (ex) {
+            let code = prompt('enter the otp', '');
+            if (code == null) return;
+            ex.confirm(code).then(function (result) {
+                Axios.post("http://localhost:3001/login", {
+                    NID, phone
+                }).then((response) => {
+                    if (response.data.message) {
+                        setLoginErr(response.data.message);
+                    } else {
+                        localStorage.setItem("loginStat", "yes");
+                        localStorage.setItem("loginCred", JSON.stringify(response.data[0]));
+                        console.log(response.data[0]);
+                        window.location.pathname="/status"
+                    }
+                });
+            }).catch((error)=>{
+                console.log(error);
+            }) 
+        })
     };
     return (
         <>
@@ -42,7 +54,7 @@ export default function Login() {
                             setPhone(e.target.value);
                         }} />
                 </div>
-                <h3 className="alert">{loginErr}</h3>
+                <h3 className="alert" id="recaptcha">{loginErr}</h3>
                 <input type="submit" value="Login" />
             </form>
         </>
